[SDESK-7994] Split the `superdesk.ai` services into a file each
Review follow-ups on the module layout. The package `__init__` no longer imports `module`, so `MODULES` names `superdesk.ai.module` and importing `superdesk.ai.<x>` no longer pulls the resource configs in with it. The `TYPE_CHECKING` guard in `providers` stays: `models` imports `allowed_provider_types`, so the cycle it avoids is real and unrelated to the package import.

Also drops the redundant `@dataclass` on `AIActionParameters`, which the `Dataclass` base already applies, reads the run payload with `from_dict` to match the rest of the codebase, and documents `AI_REQUEST_TIMEOUT` against the server and proxy timeouts it has to sit under.
